src/test/javascript/spec/app/entities/dialogs/dialogs-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';

import AlertService from '@/shared/alert/alert.service';
import * as config from '@/shared/config/config';
import DialogsUpdateComponent from '@/entities/dialogs/dialogs-update.vue';
import DialogsClass from '@/entities/dialogs/dialogs-update.component';
import DialogsService from '@/entities/dialogs/dialogs.service';

import CompetitorsService from '@/entities/competitors/competitors.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const i18n = config.initI18N(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.component('font-awesome-icon', {});

describe('Component Tests', () => {
  describe('Dialogs Management Update Component', () => {
    let wrapper: Wrapper<DialogsClass>;
    let comp: DialogsClass;
    let dialogsServiceStub: SinonStubbedInstance<DialogsService>;

    beforeEach(() => {
      dialogsServiceStub = sinon.createStubInstance<DialogsService>(DialogsService);

      wrapper = shallowMount<DialogsClass>(DialogsUpdateComponent, {
        store,
        i18n,
        localVue,
        router,
        provide: {
          alertService: () => new AlertService(store),
          dialogsService: () => dialogsServiceStub,

          competitorsService: () => new CompetitorsService()
        }
      });
      comp = wrapper.vm;
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: 123 };
        comp.dialogs = entity;
        dialogsServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(dialogsServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.dialogs = entity;
        dialogsServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(dialogsServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });
  });
});
